requirements.txt
django-oscar==3.1